requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.43
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
frozenlist==1.7.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
greenlet==3.2.4
gspread==6.2.1
idna==3.10
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.4
oauthlib==3.3.1
propcache==0.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
SQLAlchemy[asyncio]==2.0.43
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
